feat : 로그인한 사용자만 수정, 선택 삭제 가능

Closes #47

See merge request cloud_track/class_04/web_project2/team06/Shoux-Kream!51

+++ src/main/java/com/shoux_kream/cart/controller/CartApiContorller.java
@@ -3,9 +3,14 @@ package com.shoux_kream.cart.controller;
 import com.shoux_kream.cart.dto.CartRequestDto;
 import com.shoux_kream.cart.dto.CartResponseDto;
 import com.shoux_kream.cart.service.CartService;
+import com.shoux_kream.user.dto.response.UserResponse;
+import com.shoux_kream.user.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.User;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,6 +21,7 @@ import java.util.List;
 public class CartApiContorller {
 
     private final CartService cartService;
+    private final UserService userService;
 
     // 장바구니 담기
     @PostMapping("/add/{itemId}")
@@ -28,29 +34,50 @@ public class CartApiContorller {
     }
 
     // 장바구니 조회
-    @GetMapping("/summary/{userId}")
-    public ResponseEntity<List<CartResponseDto>> allCarts(@PathVariable("userId") Long userId) {
-        List<CartResponseDto> carts = cartService.allCarts(userId);
+    @GetMapping("/summary")
+    public ResponseEntity<List<CartResponseDto>> allCarts(@AuthenticationPrincipal User principal) {
+        if (principal == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        String email = principal.getUsername();
+        UserResponse userResponse = userService.getUser(email);
+        List<CartResponseDto> carts = cartService.allCarts(userResponse.getUserId());
 
-        return ResponseEntity.ok()
-                .body(carts);
+        return ResponseEntity.ok(carts);
     }
 
     // 장바구니 수정 -> 장바구니에서 수량 및 옵션 수정으로 사용
     @PatchMapping("/edit/{cartId}")
-    public ResponseEntity<CartResponseDto> updateCart(@Valid @RequestBody CartRequestDto cartRequestDto, @PathVariable("cartId") Long cartId) {
+    public ResponseEntity updateCart(@Valid @RequestBody CartRequestDto cartRequestDto, @PathVariable("cartId") Long cartId, @AuthenticationPrincipal User principal) {
+        if (principal == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        String email = principal.getUsername();
+
+        if(!cartService.validateCartItem(cartId, email)){
+            return new ResponseEntity<String>("수정 권한이 없습니다.", HttpStatus.FORBIDDEN);
+        }
+
         CartResponseDto cart = cartService.updateCart(cartRequestDto, cartId);
 
-        return ResponseEntity.ok()
-                .body(cart);
+        return ResponseEntity.ok(cart);
 
     }
     
     // 장바구니 삭제
     // 장바구니 일괄 삭제
     @DeleteMapping("/")
-    public ResponseEntity<Void> deleteCarts(@RequestBody List<Long> cartIds) {
-        cartService.deleteAllCarts(cartIds);
+    public ResponseEntity deleteCarts(@RequestBody List<Long> cartIds, @AuthenticationPrincipal User principal) {
+        if (principal == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        String email = principal.getUsername();
+
+        for (Long cartId : cartIds) {
+            cartService.validateCartItem(cartId, email);
+        }
+
+        cartService.deleteCarts(cartIds);
 
         return ResponseEntity.ok()
                 .build();
@@ -58,7 +85,16 @@ public class CartApiContorller {
     
     // 장바구니 개별 삭제
     @DeleteMapping("/{cartId}")
-    public ResponseEntity<Void> deleteCart(@PathVariable("cartId") Long cartId) {
+    public ResponseEntity deleteCart(@PathVariable("cartId") Long cartId, @AuthenticationPrincipal User principal) {
+        if (principal == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        String email = principal.getUsername();
+
+        if(!cartService.validateCartItem(cartId, email)){
+            return new ResponseEntity<String>("삭제 권한이 없습니다.", HttpStatus.FORBIDDEN);
+        }
+
         cartService.deleteCart(cartId);
 
         return ResponseEntity.ok()
